Fix test setup and remove render hack
