+++ .travis.yml
@@ -16,7 +16,7 @@ env:
 
 matrix:
   include:
-    - php: 7.1
+    - php: 7.2
       env:
         - COMPOSER_FLAGS="--optimize-autoloader --prefer-dist --no-interaction --prefer-lowest --prefer-stable"
     - php: 7.3
