+++ Resources/config/routing_api.yaml
@@ -1,3 +1,3 @@
 sulu_community.blacklist_item:
     type: rest
-    resource: SuluCommunityBundle/Controller/BlacklistItemController.php
+    resource: Sulu\Bundle\CommunityBundle\Controller\BlacklistItemController
