+++ Tests/Application/.gitattributes
@@ -1,0 +1,11 @@
+.gitattributes export-ignore
+.gitignore export-ignore
+.github export-ignore
+.styleci.yml export-ignore
+.scrutinizer.yml export-ignore
+.travis.yml export-ignore
+.php_cs.dist export-ignore
+appveyor.yml export-ignore
+phpstan.neon export-ignore
+phpunit.xml.dist export-ignore
+/Tests export-ignore
