+++ Tests/Application/config/config.yml
@@ -1,0 +1,19 @@
+sulu_community:
+    webspaces:
+        sulu-io:
+            from: 'admin@localhost'
+
+doctrine:
+    orm:
+        mappings:
+            gedmo_tree:
+                type: xml
+                prefix: Gedmo\Tree\Entity
+                dir: "%kernel.root_dir%/../../vendor/gedmo/doctrine-extensions/lib/Gedmo/Tree/Entity"
+                alias: GedmoTree
+                is_bundle: false
+
+sulu_media:
+    ffmpeg:
+        ffmpeg_binary: null
+        ffprobe_binary: null
