+++ Tests/Application/config/config_admin.yml
@@ -1,20 +1,11 @@
+imports:
+    - config.yml
+
 parameters:
     secret: test
 
 framework:
     router: { resource: "%kernel.root_dir%/config/routing_admin.yml" }
 
-sulu_community:
-    webspaces:
-        sulu-io:
-            from: 'admin@sulu.io'
-
-doctrine:
-    orm:
-        mappings:
-            gedmo_tree:
-                type: xml
-                prefix: Gedmo\Tree\Entity
-                dir: "%kernel.root_dir%/../../vendor/gedmo/doctrine-extensions/lib/Gedmo/Tree/Entity"
-                alias: GedmoTree
-                is_bundle: false
+swiftmailer:
+    disable_delivery:  true
